# swagger.json
{
    "swagger": "2.0",
    "info": {
      "description": "This is a forum about technologie, all minds are welcome. Also is a project from a Jedi and his Padawan.",
      "version": "1.0.0",
      "title": "TechLife Forum",
      "contact": {
        "email": "juanfbl9307@gmail.com"
      }
    },
    "host": "localhost:3000",
    "basePath": "/",
...
    "schemes": [
      "http"
    ],
    "paths": {
      "/user": {
        "post": {
          "tags": [
            "user"
          ],
          "summary": "Create user",
          "description": "Anyone can sign in all will be created as Poster",
          "operationId": "UserController.createPoster",
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "body",
              "name": "body",
              "description": "Created user object",
              "required": true,
              "schema": {
                "$ref": "#/definitions/User"
              }
            }
          ],
          "responses": {
            "default": {
              "description": "successful operation"
            }
          }
        }
      },
...
    },
    "securityDefinitions": {
      "petstore_auth": {
        "type": "oauth2",
        "authorizationUrl": "http://petstore.swagger.io/oauth/dialog",
        "flow": "implicit",
        "scopes": {
          "write:pets": "modify pets in your account",
          "read:pets": "read your pets"
        }
      },
      "api_key": {
        "type": "apiKey",
        "name": "api_key",
        "in": "header"
      }
    },
    "definitions": {
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      }
    },
...
  }